Use the Twitter gem to check for username
After much gnashing of teeth, I have decided that the best way to check
for the availability of a Twitter username is to use the damn gem with
valid credentials.

I do this under protest - it's ridiculous that you have to create a
Twitter application simply to check to see if a username is available.

Still, this is the safest, most accurate way to check.