Enqueue jobs in controller
I FINALLY figured out some issues I've been having about jobs not
working correctly. What would happen is that an
`ActiveRecord::RecordNotFound` error would get thrown and I couldn't
figure out why. Turns out that `after_create` is still inside a
transaction and is the wrong place to enqueue a job.

This commit moves things around so that the jobs are enqueued up at the
controller-level and that's been much more stable.

The only problem with this commit is that integration tests are now
failing and I haven't been able to figure out why.